backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
channels>=4.0.0
daphne>=4.0.0
psycopg2-binary>=2.9.9
Pillow>=10.0.0
python-dotenv>=1.0.0
django-storages>=1.14.2
boto3>=1.34.0
channels-redis>=4.1.0
django-filter>=23.5
django-cleanup>=8.0.0

# Nuove dipendenze
djangorestframework-simplejwt>=5.3.1
django-redis>=5.4.0
redis>=5.0.1
django-ratelimit>=4.1.0 